feat: Implement Card Purchase, Reward System, and Monster Death Tracking
- Implement card purchase functionality:
  - Add `buyCard()` function in `CardViewModel` to handle card purchases.
  - Update `CardDao` to increment `numberCardCount` when a card is purchased.
  - Introduce `updateCard()` to mark a card as discovered and update its effect.
- Enhance reward system:
  - Add `RewardViewModel` to manage reward state.
  - Implement `RewardRepository` and `RewardDao` for database interactions.
  - Add `insertReward()` to initialize rewards in the database.
  - Add `updatePurchaseCost()`, `updateGoldBut()`, `updateTotalDeath()`, `updateGoldMonsterDeath()` to manage rewards.
 - Update Monster death tracking:
  - Modify `MonsterDao` to track `deathCount` using `BigInteger`.
  - Update `updateMonster()` in `MonsterDao` to use `BigInteger` for `deathCount`.
  - Update `Monster` model to use `BigInteger` for `deathCount`.
  - Update `MonsterViewModel` to update the reward SQLite database.
 - Refactor Card and Reward models to use `BigDecimal` and `BigInteger` for numeric values.
 - Add `Converters` class to handle `BigDecimal` and `BigInteger` in `AppDatabase`.
- Add error handling and default values:
  - Implement `loadCards()` and `loadMonster()` to recover if there is an error and no default data.
  - Add generic `HeroCard` to handle null cards.
- Update UI components:
  -  Update `MonsterZone` to display new reward values.
  - Update `HeroesZone` to display a list of 6 cards, using the generic card if the list is empty.
 - Add `loadAllCards()` in `CardViewModel` to load all cards in the database if they do not exist.
 - Add `loadReward()` in `RewardViewModel` to load the reward in the database.

+++ app/src/main/java/com/example/beatfranticallyidle/data/source/local/card/CardEntity.kt
@@ -1,5 +1,6 @@
 package com.example.beatfranticallyidle.data.source.local.card
 
+import android.icu.math.BigDecimal
 import androidx.annotation.DrawableRes
 import androidx.room.Embedded
 import androidx.room.Entity
@@ -12,6 +13,7 @@ import com.example.beatfranticallyidle.data.source.local.card.model.CardEffect
 import com.example.beatfranticallyidle.data.source.local.card.model.CardElement
 import com.example.beatfranticallyidle.data.source.local.card.model.CardType
 import com.example.beatfranticallyidle.data.source.local.card.model.CardWithCardType
+import java.math.BigInteger
 
 @Entity(tableName = "card_elements")
 data class CardTypeEntity(
@@ -47,7 +49,7 @@ data class CardEntity(
     val effectDescription: String,
     val discovered: Boolean,
     val effectActivated: Boolean,
-    val numberCardCount: Int,
+    val numberCardCount: BigInteger,
     val imageTypeResId: Int
 ) {
     fun toCard(): Card {
